Vježbe/NoveVjezbe/CountryCodes/requirements.txt
sv_ttk
Pillow